requirements.txt
discord.py
pillow
aiohttp
python-dotenv
readability-lxml
beautifulsoup4
aiosqlite
